pyproject.toml
[project]
name = "audiblez"
version = "0.4.10"
description = "Generate audiobooks from e-books (epub to wav/m4b)"
authors = [
    { name = "Claudio Santini", email = "hireclaudio@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "ebooklib (>=0.18,<0.19)",
    "soundfile (>=0.13.1,<0.14.0)",
    "pick (>=2.4.0,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "spacy (>=3.8.3,<4.0.0)",
    "kokoro (>=0.9.4,<0.10.0)",
    "misaki[zh] (>=0.9.4,<0.10.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "wxpython>=4.2.3",
    "pillow>=11.3.0",
    "pip>=25.2",
    "pyyaml>=6.0.0",
]
exclude = [
    "imgs",
    "test",
    "*.wav",
    "*.m4b",
    "cover",
    "chapters.txt"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://claudio.uk/posts/epub-to-audiobook.html"
Documentation = "https://github.com/santinic/audiblez"
Repository = "https://github.com/santinic/audiblez"
Issues = "https://github.com/santinic/audiblez/issues"

[project.scripts]
audiblez = "audiblez.cli:cli_main"
audiblez-ui = "audiblez.ui:main"

[dependency-groups]
dev = [
    "deptry>=0.23.0",
]
